refactor: most of the project
Change configuration parameters:
    Server.Port cannot be a range anymore
    Testing.StartServer -> Testing.RunServer
    NEW Testing.PortRange will be used for parallel tests
    REMOVE Client.AccountAliases and related feature
    REMOVE Client.GasLogDir, Client.EnableGasLog and related feature, to
     be re-introduced with a Project parameter in the future
    Client.ContractBuildDir and Compiler.BuildDir -> Project.ObjectDir
    Client.DefaultGas -> Client.GasLimit
    NEW Client.GasPrice
    Compiler.ContractDir -> Project.SourceDir
    Compiler.Lint.* -> Linter.*

Change Factory behavior:
    create_client does not add compiled contracts to the client, this
     must be done explicitly with client.update_contracts(...) now.
     Factory provides new methods get_sourcelist and get_objectlist

setup.py
    move CLI entry points

solitude/_commandline/main.py
solitude/_commandline/cmd_compile.py
solitude/_commandline/cmd_debug.py
solitude/_commandline/cmd_init.py
solitude/_commandline/cmd_install.py
solitude/_commandline/cmd_lint.py
solitude/_commandline/cmd_server.py
solitude/_commandline/cmd_trace.py
solitude/_commandline/text_util.py
solitude/_commandline/color_util.py
    move each subcommand to its own file. main imports modules as needed
     when executing a subcommand.

solitude/_internal/oi_interface.py
solitude/_internal/oi_common_objects.py
solitude/_internal/oi_serializable.py
    move the 'object interface' base and class ColorText to
     _internal/oi_*.

solitude/_internal/config_schema.py
    change some configuration parameters as described above

tests/conftest.py
tests/install_tools_locally.py
    add possibility to use locally installed tools with
     --local-tools, instead of downloading tools every time. provide
     script to install tools locally to the location expected by the
     tests when --local-tools is enabled (~/.solitude-dev/testing/)

solitude/errors.py
solitude/_internal/errors.py
solitude/common/errors.py
    regroup exceptions. remove exceptions that were too narrow or
     unnecessarily specific to one of the tools.
    hide CLIError and InternalError
    RPCError -> CommunicationError
    AccountError -> ValueError
    InstallError -> CommunicationError
    ToolNotFoundError -> SetupError
    REMOVE CallForbiddenError

solitude/client/eth_client.py
    Cannot pass compiled contracts in constructor any more. must call
     update_contracts(...)

solitude/debugger/*
    move debugger to its own directory

solitude/factory.py
    move factory to the root because we don't want anything inside
     'common' to depend on all other submodules

solitude/linter/*
    move linter to its own directory

and more refactoring towards the following dependency constraints
_cli -> debugger -> | client  ->|
                    | server  ->| tools -> common -> _internal
        testing  -> | linter  ->|
                    | compiler->|
