+++ solitude/common/__init__.py
@@ -8,17 +8,32 @@ from solitude.common.structures import (
     file_message_format,
     TransactionInfo,
     hex_repr)
+from solitude.common.resource_util import (
+    get_resource_path, get_global_config, update_global_config, copy_from_url, read_from_url, open_url)
+from solitude.common.config_util import (
+    read_config_file, read_yaml_or_json, make_default_config)
 
 from solitude.common.contract_objectlist import ContractObjectList
 from solitude.common.contract_sourcelist import ContractSourceList
-
 from solitude.common.dump import Dump
+
 
 __all__ = [
     "FileMessage",
     "file_message_format",
     "TransactionInfo",
     "hex_repr",
+
+    "get_resource_path",
+    "get_global_config",
+    "update_global_config",
+    "copy_from_url",
+    "read_from_url",
+    "open_url",
+
+    "read_config_file",
+    "read_yaml_or_json",
+    "make_default_config",
 
     "ContractObjectList",
     "ContractSourceList",
