+++ solitude/common/resource_util.py
@@ -9,12 +9,14 @@ import shutil
 import io
 import requests
 import contextlib
+import solitude._internal
 from solitude._internal import RaiseForParam, type_assert, EnumType, value_assert
+from solitude.common.errors import CommunicationError
 
 
 def get_resource_path(resource_name):
     return os.path.join(
-        os.path.dirname(os.path.abspath(__file__)),
+        os.path.dirname(os.path.abspath(solitude._internal.__file__)),
         "resources",
         resource_name)
 
@@ -33,10 +35,10 @@ def _parse_url(url):
     return (Schema.FILE, url)
 
 
-def _copy_fileobj_to_destination(source, destination):
+def _copy_fileobj_to_destination(source, destination, is_text=False):
     with contextlib.ExitStack() as stack:
         if isinstance(destination, str):
-            fp = stack.enter_context(open(destination, "wb"))
+            fp = stack.enter_context(open(destination, "wb" if not is_text else "w"))
         elif hasattr(destination, "write"):
             fp = destination
         else:
@@ -45,27 +47,44 @@ def _copy_fileobj_to_destination(source, destination):
         shutil.copyfileobj(source, fp)
 
 
-def _url_to_fileobj(url: str):
+def _url_to_fileobj(url: str, decode=False):
     schema, path = _parse_url(url)
-    if schema in (Schema.HTTP, Schema.HTTPS):
-        r = requests.get(url, stream=True)
-        r.raise_for_status()
-        r.raw.decode_content = True
-        yield r.raw
-    elif schema == Schema.FILE:
-        with open(path, "rb") as fp:
-            yield fp
-    elif schema == Schema.RESOURCE:
-        with open(get_resource_path(path), "rb") as fp:
-            yield fp
-    else:
-        value_assert(False, "Schema not recognized for URL: %s" % url)
+    try:
+        if schema in (Schema.HTTP, Schema.HTTPS):
+            r = requests.get(url, stream=True)
+            r.raise_for_status()
+            r.raw.decode_content = True
+            stream = r.raw
+            if decode:
+                stream = io.TextIOWrapper(r.raw)
+            return stream
+        elif schema == Schema.FILE:
+            return open(os.path.expanduser(path), "rb" if not decode else "r")
+        elif schema == Schema.RESOURCE:
+            return open(get_resource_path(path), "rb" if not decode else "r")
+    except (requests.RequestException, FileNotFoundError, IOError) as e:
+        raise CommunicationError("Cannot open '%s'" % url) from e
+    raise CommunicationError("Schema not recognized for '%s'" % url)
 
 
-def copy_from_url(url: str, destination):
-    source = _url_to_fileobj(url)
-    _copy_fileobj_to_destination(next(source), destination)
-    _ = list(source)
+def open_url(url: str, decode=False):
+    return _url_to_fileobj(url, decode=decode)
+
+
+def read_from_url(url: str, decode=False):
+    with _url_to_fileobj(url, decode=decode) as source:
+        if decode:
+            data = io.StringIO()
+        else:
+            data = io.BytesIO()
+        _copy_fileobj_to_destination(source, data, is_text=decode)
+        data.seek(0, 0)
+        return data.read()
+
+
+def copy_from_url(url: str, destination, decode=False):
+    with _url_to_fileobj(url, decode=decode) as source:
+        _copy_fileobj_to_destination(source, destination, is_text=decode)
 
 
 class _GlobalConfig:
@@ -80,7 +99,7 @@ class _GlobalConfig:
         return self._config
 
 
-def _DEBUG_SolcEmscripten(enable):
+def _DEBUG_SolcEmscripten(enable_emscripten):
     from solitude.tools import solc
     with RaiseForParam("enable_emscripten"):
         type_assert(enable_emscripten, bool)
