+++ solitude/debugger/__init__.py
@@ -1,0 +1,14 @@
+# Copyright (c) 2019, Solitude Developers
+#
+# This source code is licensed under the BSD-3-Clause license found in the
+# COPYING file in the root directory of this source tree
+
+from solitude.debugger.evm_trace import EvmTrace, TraceStep, SourceMapping, CallStackElement, CallStackEvent
+from solitude.debugger.evm_debug_core import EvmDebugCore, Function, Frame, Step, Variable
+from solitude.debugger.oi_debugger import InteractiveDebuggerOI
+
+__all__ = [
+    "EvmTrace", "TraceStep", "SourceMapping", "CallStackElement", "CallStackEvent",
+    "EvmDebugCore", "Function", "Frame", "Step", "Variable",
+    "InteractiveDebuggerOI"
+]
