+++ solitude/linter/__init__.py
@@ -3,8 +3,8 @@
 # This source code is licensed under the BSD-3-Clause license found in the
 # COPYING file in the root directory of this source tree
 
-from solitude.compiler.compiler import Compiler
+from solitude.compiler.linter import Linter
 
 __all__ = [
-    "Compiler"
+    "Linter"
 ]
